requirements.txt
annotated-types==0.7.0
anyio==4.9.0
backoff==1.11.1
certifi==2025.6.15
click==8.2.1
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
joblib==1.5.1
multidict==6.4.4
nltk==3.9.1
numpy==2.3.1
openai==1.90.0
pydantic==2.11.7
pydantic-core==2.33.2
python-dotenv==1.1.0
regex==2024.11.6
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.14.0
typing-inspection==0.4.1

defusedxml~=0.7.1
pillow~=11.3.0
colorama~=0.4.6
requests~=2.32.4
pytz~=2025.2
attrs~=25.3.0
PyYAML~=6.0.2
propcache~=0.3.2
typing_extensions~=4.14.0
Jinja2~=3.1.6
banks~=2.2.0
griffe~=1.9.0
platformdirs~=4.3.8
charset-normalizer~=3.4.2
setuptools~=80.9.0
fsspec~=2025.7.0
aiohttp~=3.12.15
yarl~=1.20.1
urllib3~=2.5.0
packaging~=25.0
MarkupSafe~=3.0.2
pandas~=2.2.3
pydantic_core~=2.33.2
Deprecated~=1.2.18
aiosignal~=1.4.0
frozenlist~=1.7.0
aiohappyeyeballs~=2.6.1
greenlet~=3.2.3
SQLAlchemy~=2.0.42
six~=1.17.0
python-dateutil~=2.9.0.post0
tenacity~=9.1.2
filetype~=1.2.0
networkx~=3.5
tiktoken~=0.9.0
aiosqlite~=0.21.0
dirtyjson~=1.0.8
wrapt~=1.17.2
llama-index-core~=0.14.3
dataclasses-json~=0.6.7
beautifulsoup4~=4.13.4
marshmallow~=3.26.1
typing-inspect~=0.9.0
mypy_extensions~=1.1.0

llama-index-core~=0.13.0
tiktoken~=0.9.0

rsa~=4.9.1
pyasn1~=0.6.1
pytest~=8.4.2
ollama~=0.6.0
Pygments~=2.19.2
typing_extensions~=4.14.0
pyasn1_modules~=0.4.2
cachetools~=5.5.2
websockets~=15.0.1
pydantic_core~=2.33.2
pluggy~=1.6.0
iniconfig~=2.1.0
filelock~=3.19.1
tokenizers~=0.22.1
transformers~=4.57.0
huggingface-hub~=0.35.3
safetensors~=0.6.2
mypy_extensions~=1.1.0
grpcio~=1.75.1
protobuf~=5.29.5
pyparsing~=3.2.5
typing_extensions~=4.14.0
pyasn1_modules~=0.4.2
google-api-python-client~=2.184.0
pydantic_core~=2.33.2
mypy_extensions~=1.1.0